requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pg-embedded>=0.10.0  # Embedded PostgreSQL for development

# AI/ML
openai>=1.0.0

# Utilities
psutil>=5.9.0  # For process monitoring
requests>=2.31.0
python-multipart>=0.0.6

# Optional: For vector embeddings (if using pgvector)
# pgvector>=0.2.0